server/requirements.txt
matplotlib
fastapi
pandas
pydantic
openai
tabulate
uvicorn
python-dotenv
pymongo